+++ src/comfystream/server/utils/__init__.py
@@ -1,2 +1,2 @@
-from .utils import patch_loop_datagram, add_prefix_to_app_routes, temporary_log_level, ComfyStreamTimeoutFilter
+from .utils import patch_loop_datagram, add_prefix_to_app_routes, temporary_log_level
 from .fps_meter import FPSMeter
